requirements.txt
python-binance>=1.0.34
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
streamlit>=1.30.0
streamlit-autorefresh
